Content cloner.
Cloning of content through sockets.
