+++ src/components/TranslationManager/components/TranslationContentCloner/index.ts
@@ -1,0 +1,3 @@
+import TranslationContentCloner from './TranslationContentCloner';
+
+export default TranslationContentCloner;
