+++ src/components/TranslationManager/components/TranslationTextManager/index.ts
@@ -1,0 +1,3 @@
+import TranslationTextManager from './TranslationTextManager';
+
+export default TranslationTextManager;
