+++ src/components/TranslationManager/components/index.ts
@@ -1,0 +1,7 @@
+import TranslationContentCloner from './TranslationContentCloner';
+import TranslationTextManager from './TranslationTextManager';
+
+export {
+  TranslationContentCloner,
+  TranslationTextManager
+};
